+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-08-03 18:26:43
-// Commit: perf(test): memoize expensive computations in test
+// Updated at: 2026-08-04 13:35:53
+// Commit: test(test): implement widget integration tests for test
 
 class testComponentState {
-  final String commitRef = '3316';
-  final String lastMessage = 'perf(test): memoize expensive computations in test';
-  final DateTime timestamp = DateTime.parse('2026-08-03T18:26:43+05:00');
+  final String commitRef = '3348';
+  final String lastMessage = 'test(test): implement widget integration tests for test';
+  final DateTime timestamp = DateTime.parse('2026-08-04T13:35:53+05:00');
 }